tickets in  typescript

+++ src/events/messageReactionAdd/tickets.ts
@@ -1,11 +1,12 @@
-const checkPerms = require('../../functions/checkPerms').default;
-const createTicket = require('../../functions/tickets/createTicket.js');
-const closeTicket = require('../../functions/tickets/closeTicket').default;
-const deleteTicket = require('../../functions/tickets/deleteTicket.js');
-const reopenTicket = require('../../functions/tickets/reopenTicket.js');
-const createVoice = require('../../functions/tickets/createVoice.js');
+import checkPerms from '../../functions/checkPerms';
+import createTicket from '../../functions/tickets/createTicket';
+import closeTicket from '../../functions/tickets/closeTicket';
+import deleteTicket from '../../functions/tickets/deleteTicket';
+import reopenTicket from '../../functions/tickets/reopenTicket';
+import createVoice from '../../functions/tickets/createVoice';
+import { Client, MessageReaction, TextChannel, User } from 'discord.js';
 
-module.exports = async (client, reaction, user) => {
+export default async (client: Client, reaction: MessageReaction, user: User) => {
 	// Check if author is a bot or guild is undefined
 	if (user.bot || !reaction.message.guildId) return;
 
@@ -13,17 +14,19 @@ module.exports = async (client, reaction, user) => {
 	const guild = await client.guilds.fetch(reaction.message.guildId);
 
 	// Check if the bot has permission to manage messages
-	const permCheck = checkPerms(['ReadMessageHistory'], guild.members.me, reaction.message.channelId);
+	const permCheck = checkPerms(['ReadMessageHistory'], guild.members.me!, reaction.message.channelId);
 	if (permCheck) return;
 
 	// Fetch the reaction's message
-	const message = await reaction.message.fetch().catch(err => logger.error(err));
+	const message = await reaction.message.fetch().catch(err => { logger.error(err); return null; });
+	if (!message) return;
 
 	// Get the reaction's emoji
 	const emojiId = reaction.emoji.id ?? reaction.emoji.name;
 
 	// Get the member
-	const member = await guild.members.fetch(user.id).catch(err => logger.error(err));
+	const member = await guild.members.fetch(user.id).catch(err => { logger.error(err); return null });
+	if (!member) return;
 
 	// Get current settings for the guild and check if tickets are enabled
 	const srvconfig = await sql.getData('settings', { guildId: guild.id });
@@ -36,25 +39,26 @@ module.exports = async (client, reaction, user) => {
 			await reaction.users.remove(member.id);
 		}
 		else if (emojiId == '⛔') {
-			await deleteTicket(client, srvconfig, member, message.channel);
+			await deleteTicket(message.channel as TextChannel);
 		}
 		else if (emojiId == '🔓') {
-			await reopenTicket(client, srvconfig, member, message.channel);
+			await reopenTicket(srvconfig, member, message.channel as TextChannel);
 			await reaction.users.remove(member.id);
 		}
 		else if (emojiId == '🔒') {
-			if (message.embeds[0] && !message.embeds[0].title.includes('Ticket Created')) return;
-			await closeTicket(srvconfig, member, message.channel);
+			if (message.embeds[0] && message.embeds[0].title !== 'Ticket Created') return;
+			await closeTicket(srvconfig, member, message.channel as TextChannel);
 			await reaction.users.remove(member.id);
 		}
 		else if (emojiId == '🔊') {
 			if (message.embeds[0] && message.embeds[0].title !== 'Ticket Created') return;
-			await createVoice(client, srvconfig, member, message.channel);
+			await createVoice(client, srvconfig, member, message.channel as TextChannel);
 			await reaction.users.remove(member.id);
 		}
 	}
 	catch (err) {
 		const msg = await error(err, message, true);
+		if (!msg) return;
 		await sleep(5000);
 		await msg.delete();
 	}
